+++ Classes/Model/Type/FundingAgency.php
@@ -55,6 +55,7 @@ final class FundingAgency extends AbstractType
         'funder',
         'funding',
         'globalLocationNumber',
+        'hasCertification',
         'hasCredential',
         'hasMerchantReturnPolicy',
         'hasOfferCatalog',