+++ Classes/Model/Type/OfferForLease.php
@@ -25,6 +25,7 @@ final class OfferForLease extends AbstractType
     protected static array $propertyNames = [
         'acceptedPaymentMethod',
         'addOn',
+        'additionalProperty',
         'additionalType',
         'advanceBookingRequirement',
         'aggregateRating',