+++ Classes/Model/Type/OfferForPurchase.php
@@ -25,6 +25,7 @@ final class OfferForPurchase extends AbstractType
     protected static array $propertyNames = [
         'acceptedPaymentMethod',
         'addOn',
+        'additionalProperty',
         'additionalType',
         'advanceBookingRequirement',
         'aggregateRating',