+++ Classes/Model/Type/PublicToilet.php
@@ -46,6 +46,7 @@ final class PublicToilet extends AbstractType
         'geoTouches',
         'geoWithin',
         'globalLocationNumber',
+        'hasCertification',
         'hasDriveThroughService',
         'hasMap',
         'identifier',